added additional tests for factorial

+++ src/main/java/PACKAGE_NAME/Factorial.java
@@ -2,9 +2,9 @@ package PACKAGE_NAME;
 
 public class Factorial {
 
-    public static Integer factorial (int n) {
+    public static Long factorial (int n) {
         if (n == 0) {
-            return 1;
+            return 1l;
         }
         return n * factorial(n - 1);
     }
